+++ pyinstaller_hooks/hook-ortools.py
@@ -1,5 +1,5 @@
 from PyInstaller.utils.hooks import collect_dynamic_libs
 
 
-# OR-Tools loads these DLLs from ``ortools/.libs`` at import time on Windows.
+# OR-Tools loads native solver libraries from ``ortools/.libs`` at import time.
 binaries = collect_dynamic_libs("ortools", destdir="ortools/.libs")